+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [2.7.1] - 2018-04-25
+
+### Changed
+- `TwigTemplateChoice` now set template basename as choice key 
+
 ## [2.7.0] - 2018-04-25
 
 ### Added